requirements.txt
keyboard
pyserial
